requirements.txt
discord.py
asyncpg
python-dotenv
apscheduler
asyncpg
authlib 
httpx 
itsdangerous
python-multipart
views